Move `ibancho` package to `multiplayer` package and rename to `api`
