+++ src/com/osudroid/multiplayer/RoomChat.kt
@@ -23,8 +23,8 @@ import com.edlplan.framework.easing.Easing
 import com.edlplan.ui.BaseAnimationListener
 import com.edlplan.ui.EasingHelper
 import com.edlplan.ui.fragment.BaseFragment
-import com.osudroid.ibancho.RoomAPI
-import com.osudroid.ibancho.data.RoomPlayer
+import com.osudroid.multiplayer.api.RoomAPI
+import com.osudroid.multiplayer.api.data.RoomPlayer
 import com.reco1l.osu.mainThread
 import com.reco1l.toolkt.android.*
 import com.reco1l.toolkt.kotlin.async
