+++ src/com/osudroid/multiplayer/RoomPlayerMenu.kt
@@ -9,8 +9,8 @@ import com.edlplan.ui.BaseAnimationListener
 import com.edlplan.ui.EasingHelper
 import com.edlplan.ui.fragment.BaseFragment
 import com.edlplan.ui.fragment.WebViewFragment
-import com.osudroid.ibancho.RoomAPI
-import com.osudroid.ibancho.data.RoomPlayer
+import com.osudroid.multiplayer.api.RoomAPI
+import com.osudroid.multiplayer.api.data.RoomPlayer
 import com.reco1l.osu.ui.MessageDialog
 import com.reco1l.toolkt.android.cornerRadius
 import com.reco1l.toolkt.android.dp
