+++ src/com/osudroid/multiplayer/RoomScene.kt
@@ -3,24 +3,24 @@ package com.osudroid.multiplayer
 import com.osudroid.BuildSettings
 import com.reco1l.andengine.*
 import com.reco1l.andengine.sprite.*
-import com.osudroid.ibancho.IPlayerEventListener
-import com.osudroid.ibancho.IRoomEventListener
-import com.osudroid.ibancho.RoomAPI
-import com.osudroid.ibancho.data.PlayerStatus
-import com.osudroid.ibancho.data.PlayerStatus.*
-import com.osudroid.ibancho.data.Room
-import com.osudroid.ibancho.data.RoomBeatmap
-import com.osudroid.ibancho.data.RoomGameplaySettings
-import com.osudroid.ibancho.data.RoomMods
-import com.osudroid.ibancho.data.RoomPlayer
-import com.osudroid.ibancho.data.RoomTeam
-import com.osudroid.ibancho.data.RoomTeam.Blue
-import com.osudroid.ibancho.data.RoomTeam.Red
-import com.osudroid.ibancho.data.TeamMode
-import com.osudroid.ibancho.data.TeamMode.HeadToHead
-import com.osudroid.ibancho.data.TeamMode.TeamVersus
-import com.osudroid.ibancho.data.WinCondition
-import com.osudroid.ibancho.data.WinCondition.*
+import com.osudroid.multiplayer.api.IPlayerEventListener
+import com.osudroid.multiplayer.api.IRoomEventListener
+import com.osudroid.multiplayer.api.RoomAPI
+import com.osudroid.multiplayer.api.data.PlayerStatus
+import com.osudroid.multiplayer.api.data.PlayerStatus.*
+import com.osudroid.multiplayer.api.data.Room
+import com.osudroid.multiplayer.api.data.RoomBeatmap
+import com.osudroid.multiplayer.api.data.RoomGameplaySettings
+import com.osudroid.multiplayer.api.data.RoomMods
+import com.osudroid.multiplayer.api.data.RoomPlayer
+import com.osudroid.multiplayer.api.data.RoomTeam
+import com.osudroid.multiplayer.api.data.RoomTeam.Blue
+import com.osudroid.multiplayer.api.data.RoomTeam.Red
+import com.osudroid.multiplayer.api.data.TeamMode
+import com.osudroid.multiplayer.api.data.TeamMode.HeadToHead
+import com.osudroid.multiplayer.api.data.TeamMode.TeamVersus
+import com.osudroid.multiplayer.api.data.WinCondition
+import com.osudroid.multiplayer.api.data.WinCondition.*
 import com.reco1l.osu.mainThread
 import com.osudroid.multiplayer.Multiplayer.isConnected
 import com.osudroid.multiplayer.Multiplayer.isRoomHost
