+++ src/com/osudroid/multiplayer/api/RoomAPI.kt
@@ -1,18 +1,18 @@
-package com.osudroid.ibancho
+package com.osudroid.multiplayer.api
 
 import com.osudroid.BuildSettings
 import com.osudroid.debug.MockSocket
-import com.osudroid.ibancho.data.PlayerStatus
-import com.osudroid.ibancho.data.Room
-import com.osudroid.ibancho.data.RoomMods
-import com.osudroid.ibancho.data.RoomStatus
-import com.osudroid.ibancho.data.RoomTeam
-import com.osudroid.ibancho.data.TeamMode
-import com.osudroid.ibancho.data.WinCondition
-import com.osudroid.ibancho.data.parseBeatmap
-import com.osudroid.ibancho.data.parseGameplaySettings
-import com.osudroid.ibancho.data.parsePlayer
-import com.osudroid.ibancho.data.parsePlayers
+import com.osudroid.multiplayer.api.data.PlayerStatus
+import com.osudroid.multiplayer.api.data.Room
+import com.osudroid.multiplayer.api.data.RoomMods
+import com.osudroid.multiplayer.api.data.RoomStatus
+import com.osudroid.multiplayer.api.data.RoomTeam
+import com.osudroid.multiplayer.api.data.TeamMode
+import com.osudroid.multiplayer.api.data.WinCondition
+import com.osudroid.multiplayer.api.data.parseBeatmap
+import com.osudroid.multiplayer.api.data.parseGameplaySettings
+import com.osudroid.multiplayer.api.data.parsePlayer
+import com.osudroid.multiplayer.api.data.parsePlayers
 import com.osudroid.multiplayer.Multiplayer
 import ru.nsu.ccfit.zuev.osu.SecurityUtils
 import io.socket.client.IO
