+++ src/com/osudroid/multiplayer/api/data/Parsing.kt
@@ -1,4 +1,4 @@
-package com.osudroid.ibancho.data
+package com.osudroid.multiplayer.api.data
 
 import org.json.JSONArray
 import org.json.JSONObject
