+++ src/com/osudroid/multiplayer/api/data/RoomPlayer.kt
@@ -1,4 +1,4 @@
-package com.osudroid.ibancho.data
+package com.osudroid.multiplayer.api.data
 
 data class RoomPlayer(
