+++ src/com/osudroid/multiplayer/api/data/RoomTeam.kt
@@ -1,4 +1,4 @@
-package com.osudroid.ibancho.data
+package com.osudroid.multiplayer.api.data
 
 /**
  * Represents the team of a room player.
